Remove redundant getSuggestionVendor util
