+++ packages/domain-picker/src/index.tsx
@@ -1,2 +1,1 @@
-export { getDomainSuggestionsVendor } from './utils';
 export { useDomainSuggestions } from './hooks';
